TorXakis.Compiler

Description

Documentation

compileFile :: FilePath -> IO (Either Error (Id, TxsDefs, Sigs VarId))

Compile a file into a TorXakis model.

compileString :: String -> IO (Either Error (Id, TxsDefs, Sigs VarId))

Compile a string into a TorXakis model.

compileUnsafe :: CompilerM a -> a

Run the compiler throwing an error if the compiler returns an Error.

compileValDefs

Arguments

:: Sigs VarId

-> [VarId]

-> Int
Next unique identifier.

-> String
String to parse.

-> CompilerM (Id, Map VarId (ValExpr VarId))

Sub-compiler for value definitions

compileVarDecls

Arguments

:: Sigs VarId

-> Int
Next unique identifier.

-> String
String to parse.

-> CompilerM (Id, [VarId])

Sub-compiler for variable declarations.

compileValExpr

Arguments

:: Sigs VarId

-> [VarId]

-> Int
Next unique identifier.

-> String
String to parse.

-> CompilerM (Id, ValExpr VarId)

Sub-compiler for value expressions.

compileBExpr

Arguments

:: Sigs VarId

-> [ChanId]

-> [VarId]

-> Int
Next unique identifier.

-> String
String to parse.

-> CompilerM (Id, BExpr)

Sub-compiler for behavior expressions.

compileOffer

Arguments

:: Sigs VarId

-> [ChanId]

-> [VarId]

-> Int
Next unique identifier.

-> String
String to parse.

-> CompilerM (Id, Set Offer)

Sub-compiler for action prefix offers.

compileLegacy :: String -> (Id, TxsDefs, Sigs VarId)

Legacy compile function, used to comply with the old interface. It should be deprecated in favor of compileFile.