TorXakis.Compiler.Data

Description

Documentation

data CompilerM a

runCompiler :: CompilerM a -> StateT St (Either Error) a

Run the compiler.

State query and manipulation

data St

State of the compiler.

newState :: St

Create a new state. The start value for identifiers will be '1000' for compatibility with the current TorXakis compiler.

getUnid :: CompilerM Int

Get the value of the unique id counter.

setUnid :: Int -> CompilerM ()

Set the unique id counter to the given value.

getNextId :: CompilerM Int

Get the next unique id, and increment the unique id counter in the compiler state.

Traversals using the CompilerM monad

traverseCatch :: (a -> CompilerM b) -> [a] -> CompilerM ([(a, Error)], [b])

Like traverse, but catch and error and continue if any CompilerM action throws an error in the process.

forCatch :: [a] -> (a -> CompilerM b) -> CompilerM ([(a, Error)], [b])

Flipped version of traverseCatch