TorXakis.Compiler.Defs.BehExprDefs

Description

Compilation functions related to TorXakis behavior expressions.

Documentation

toBExpr :: (MapsTo Text SortId mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc FuncDeclE) (Signature, Handler VarId) mm, MapsTo (Loc ChanRefE) (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm, MapsTo ProcId () mm, In (Loc FuncDeclE, Signature) (Contents mm) ~ False) => mm -> Map (Loc VarRefE) (Either VarId [(Signature, Handler VarId)]) -> BExpDecl -> CompilerM BExpr

Compile a behavior expression declaration into a behavior expression.

toOffer :: (MapsTo Text SortId mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc FuncDeclE) (Signature, Handler VarId) mm, MapsTo (Loc ChanRefE) (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm, In (Loc FuncDeclE, Signature) (Contents mm) ~ False) => mm -> Map (Loc VarRefE) (Either VarId [(Signature, Handler VarId)]) -> OfferDecl -> CompilerM Offer

Compile a list offer declarations on a channel into a list of offer declarations.

toActOffer :: (MapsTo Text SortId mm, MapsTo (Loc ChanRefE) (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc FuncDeclE) (Signature, Handler VarId) mm, MapsTo (Loc VarDeclE) SortId mm, In (Loc FuncDeclE, Signature) (Contents mm) ~ False) => mm -> Map (Loc VarRefE) (Either VarId [(Signature, Handler VarId)]) -> ActOfferDecl -> CompilerM ActOffer

Compile an action offer declaration into an action offer.