TorXakis.Compiler.Defs.FuncTable

Description

Compilation functions related to TorXakis function tables.

Documentation

adtsToFuncTable :: (MapsTo Text SortId mm, MapsTo (Loc CstrE) CstrId mm) => mm -> [ADTDecl] -> CompilerM (FuncTable VarId)

Make a function table from a list of ADT's declarations.

funcDeclsToFuncTable :: (MapsTo Text SortId mm, MapsTo (Loc FuncDeclE) FuncId mm, MapsTo FuncId FuncDefInfo mm) => mm -> [FuncDecl] -> CompilerM (FuncTable VarId)

Create a function table from a list of function declarations.

fLocToSignatureHandlers :: Map (Loc FuncDeclE) FuncId -> FuncTable VarId -> CompilerM (Map (Loc FuncDeclE) (Signature, Handler VarId))

Make a map from location of function declarations to their signature and handlers, using the given map and function table.