TorXakis.Compiler.Defs.TxsDefs

Description

Compilation functions related to TorXakis definitions.

Documentation

adtsToTxsDefs :: (MapsTo Text SortId mm, MapsTo (Loc CstrE) CstrId mm) => mm -> [ADTDecl] -> CompilerM TxsDefs

Compile a list of ADT declarations into TxsDefs.

cnectDeclsToTxsDefs :: (MapsTo Text SortId mm, MapsTo Text (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc VarDeclE) VarId mm, MapsTo (Loc FuncDeclE) (Signature, Handler VarId) mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, In (Loc FuncDeclE, Signature) (Contents mm) ~ False, In (ProcId, ()) (Contents mm) ~ False, In (Loc ChanRefE, Loc ChanDeclE) (Contents mm) ~ False) => mm -> [CnectDecl] -> CompilerM (Map CnectId CnectDef)

Compile a list of connect declarations into a map from connect id's to connect definitions.

mapperDeclsToTxsDefs :: (MapsTo Text SortId mm, MapsTo Text (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc FuncDeclE) (Signature, Handler VarId) mm, MapsTo ProcId ProcDef mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc VarDeclE) VarId mm, In (Loc ChanRefE, Loc ChanDeclE) (Contents mm) ~ False, In (Loc FuncDeclE, Signature) (Contents mm) ~ False, In (ProcId, ()) (Contents mm) ~ False) => mm -> [MapperDecl] -> CompilerM (Map MapperId MapperDef)

Compile a list of mapper declarations into a map from mapper id's to mapper definitions.

modelDeclsToTxsDefs :: (MapsTo Text SortId mm, MapsTo Text (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc FuncDeclE) (Signature, Handler VarId) mm, MapsTo ProcId ProcDef mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc VarDeclE) VarId mm, In (Loc FuncDeclE, Signature) (Contents mm) ~ False, In (Loc ChanRefE, Loc ChanDeclE) (Contents mm) ~ False, In (ProcId, ()) (Contents mm) ~ False) => mm -> [ModelDecl] -> CompilerM (Map ModelId ModelDef)

Compile a list of model declarations into a map from model id's to model definitions.

purpDeclsToTxsDefs :: (MapsTo Text SortId mm, MapsTo Text (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc FuncDeclE) (Signature, Handler VarId) mm, MapsTo ProcId ProcDef mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc VarDeclE) VarId mm, In (ProcId, ()) (Contents mm) ~ False, In (Loc FuncDeclE, Signature) (Contents mm) ~ False, In (Loc ChanRefE, Loc ChanDeclE) (Contents mm) ~ False) => mm -> [PurpDecl] -> CompilerM (Map PurpId PurpDef)

Compile a list of purpose declarations into a map from purpose id's to purpose definitions.