TorXakis.Compiler.Error

Description

Compiler error type and associated functions.

Error type and field accessors.

data Error

Simple error type.

Constructors

Error
Single error.

Fields

_errorType :: ErrorType

_errorLoc :: ErrorLoc

_errorMsg :: Text

Errors [Error]
Multiple errors.

Lenses for error fields.

errorLoc :: Traversal' Error ErrorLoc

errorMsg :: Traversal' Error Text

errorType :: Traversal' Error ErrorType

Type of errors.

data ErrorType

Type of errors that can occur when compiling a TorXakis model file.

Constructors

ParseError

TypeMismatch

Undefined Entity
An entity was not defined.

MultipleDefinitions Entity
Multiple definitions for the same entity.

Unresolved Entity
An entity could not be resolved (based on the type information for instance).

Ambiguous Entity
An entity has multiple candidates (based on the type information for instance)

Missing Entity
An entity is missing from the declaration.

NoDefinition
No definition found for function or process.

InvalidExpression

CompilerPanic
An error in the compiler has happened.

data Entity

Entity to which the error is related.

Constructors

Function

Process

InitialState

State

Variable

Sort

Entity
In case an error must be generated from a generic function that cannot have access to the entity type. The error type can be made more specific by the caller of such a generic function (see for instance getUniqueElement).

Error location support.

data ErrorLoc

Location of an error.

Constructors

NoErrorLoc
The error location could not be determined. By adding location to TorXakis core types like SortId it is possible to augment the number of places in which errors can be associated to locations.

ErrorPredef Text
The error is related to a pre-defined entity.

ErrorLoc

Fields

errorLine :: Int

errorColumn :: Int

class HasErrorLoc l where

Entities that have an error location.

Minimal complete definition

getErrorLoc

Methods

getErrorLoc :: l -> ErrorLoc

getErrorLoc :: HasErrorLoc l => l -> ErrorLoc