TorXakis.Compiler.Maps

Description

Compiler specific maps and operations on them.

Map lookups

(.@) :: (HasErrorLoc k, MapsTo k v mm, Ord k, Show k, Typeable k, Typeable v) => mm -> k -> Either Error v

Infix version of lookup, for those keys that have an error location.

(.@!!) :: (HasErrorLoc l, MapsTo k v mm, Ord k, Show k, Typeable k, Typeable v) => mm -> (k, l) -> CompilerM v

Lookup the given key, and use the given location for error reporting.

(.@@) :: (HasErrorLoc k, MapsTo k v mm, Ord k, Show k, Typeable k, Typeable v) => mm -> k -> CompilerM v

Monadic version of .@.

(<!!>) :: HasErrorLoc l => CompilerM a -> l -> CompilerM a

Set the error location (monadic version).

Specific lookups

findFuncDecl :: MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm => mm -> Loc VarRefE -> Either Error [Loc FuncDeclE]

Find the nullary function declaration that corresponds to a variable reference. A nullary function is a funciton without arguments.

determineF

Arguments

:: MapsTo (Loc FuncDeclE) Signature mm

=> mm

-> [Loc FuncDeclE]

-> [SortId]
Arguments SortId.

-> Maybe SortId
Return Sort, if known.

-> [Loc FuncDeclE]

Select the function definitions that matches the given arguments and return types.

getUniqueElement :: Show a => [a] -> Either Error a

Get the function definition if the given list is a singleton, return an error otherwise.

lookupChId :: (MapsTo (Loc ChanRefE) (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm) => mm -> Loc ChanRefE -> CompilerM ChanId

Lookup the ChanId associated to the given location.

chRefsToIds :: (MapsTo (Loc ChanRefE) (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm) => mm -> [ChanRef] -> CompilerM [ChanId]

Given a list of channel references and a composite map, get the ChanId's associated to those references.

findSortId :: MapsTo Text SortId mm => mm -> (Text, Loc t) -> Either Error SortId

Lookup the SortId associated to the given name, using the location for error reporting.

findFuncSortIds :: MapsTo (Loc FuncDeclE) Signature mm => mm -> [Loc FuncDeclE] -> Either Error [SortId]

Find all the return SortId's of the given function declarations.

findSortIdM :: MapsTo Text SortId mm => mm -> (Text, Loc t) -> CompilerM SortId

Monadic version of findSortIdM

determineSH

Arguments

:: [(Signature, Handler VarId)]

-> [SortId]
SortIds of the arguments.

-> Maybe SortId
Expected return SortId (if known).

-> Either Error (Signature, Handler VarId)

Determine the signature and the handler based on the sort arguments and the expected return type (if any).

findRight :: forall a b. (Typeable a, Typeable b) => Map (Loc VarRefE) (Either a b) -> Loc VarRefE -> Either Error b

Find the right element (if any) that corresponds to the given location.

usedChIdMap :: (MapsTo (Loc ChanRefE) (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm) => mm -> Map (Loc ChanRefE) ChanId

Channels referred in the model

findVarIdM :: (MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc VarDeclE) VarId mm) => mm -> Loc VarRefE -> CompilerM VarId

Monadic version of findVarId.

idefsNames :: MapsTo (Loc FuncDeclE) FuncId mm => mm -> [Text]

Extract the names of the implicit function declarations in the map.

usedChIds :: (MapsTo (Loc ChanRefE) (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm) => mm -> [Set ChanId]

Get the list of channel ids used in an expression (model, purpose, etc).

Map manipulation

join :: (Ord k, Ord v) => Map k v -> Map v w -> Map k w

Join two maps m0 and m1. A pair '(k, w)' is in the resulting map if and only if there is a key v such that '(k, v)' is in m0 and '(v, w)' is in m1.

dropHandler :: Map (Loc FuncDeclE) (Signature, Handler VarId) -> Map (Loc FuncDeclE) Signature

Drop the handler from the map.