TorXakis.Compiler.Maps.DefinesAMap

Description

Expressions that define a map.

Documentation

class (Show k, Ord k, Eq k) => DefinesAMap k v e mm where

Abstract syntax tree types that define a map.

'DefinesAMap k v e mm' models the fact that an AST e defines a map from k to v, provided a map mm is available.

Minimal complete definition

uGetKVs

Methods

getKVs :: mm -> e -> CompilerM [(k, v)]

Get the k v pairs defined by the ast e.

getMap :: Ord k => mm -> e -> CompilerM (Map k v)

Get the dictionary of k v pairs defined by the ast e.

getKs :: e -> [k]

Get the list of keys defined in e. This is used for invariant checking.

getKs :: (Data e, Typeable k) => e -> [k]

Get the list of keys defined in e. This is used for invariant checking.

uGetKVs :: mm -> e -> CompilerM [(k, v)]

Get the k v pairs defined by the ast e, without checking for the invariant. Unchecked version of getKVs.

getKVs :: DefinesAMap k v e mm => mm -> e -> CompilerM [(k, v)]

Get the k v pairs defined by the ast e.

getKs :: DefinesAMap k v e mm => e -> [k]

Get the list of keys defined in e. This is used for invariant checking.

getMap :: (DefinesAMap k v e mm, Ord k) => mm -> e -> CompilerM (Map k v)

Get the dictionary of k v pairs defined by the ast e.

uGetKVs :: DefinesAMap k v e mm => mm -> e -> CompilerM [(k, v)]

Get the k v pairs defined by the ast e, without checking for the invariant. Unchecked version of getKVs.

predefChDecls :: Map Text (Loc ChanDeclE)

Predefined channels declarations.