TorXakis.Compiler.Maps.VarRef

Description

Functions to manipulate maps of variable references.

Documentation

varIdForRef :: forall a. Typeable a => Map (Loc VarRefE) (Either VarId a) -> Loc VarRefE -> Either Error VarId

Get the variable id that corresponds to the given reference.

varDefsFromExp :: (MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc VarDeclE) VarId mm, MapsTo (Loc FuncDeclE) (Signature, Handler VarId) mm, Data e) => mm -> e -> Either Error (Map (Loc VarRefE) (Either VarId [(Signature, Handler VarId)]))

Given an expression, construct a map from variable references to the concrete TorXakis entities they refer to.