TorXakis.Compiler.MapsTo

Description

Definition for the MapsTo class, and associated operators.

A constraint of the form 'MapsTo k v mm' denotes that mm contains a map from keys of type k to values of type t.

Predicates on maps

class In (k, v) (Contents mm) ~ True => MapsTo k v mm where

mm maps keys of type k onto values of type v. The type mm can be thought as a composite map, which contains multiple maps. See :& for more details.

Minimal complete definition

lookup, innerMap, (<.+>), replaceInnerMap

Methods

lookup :: (Ord k, Show k, Typeable k, Typeable v) => k -> mm -> Either Error v

Lookup a key in the map.

lookupM :: (Ord k, Show k, Typeable k, Typeable v) => k -> mm -> CompilerM v

Monadic version of lookup.

innerMap :: mm -> Map k v

Get the inner map.

(<.+>) :: Ord k => Map k v -> mm -> mm infixl 4

Add a map. The addition is biased towards the new map. If the two maps have the same key, then the value of the right map is discarded, and the value of the left map is kept.

replaceInnerMap :: mm -> Map k v -> mm

Replace the inner map with the given one.

type family In (x :: *) (ys :: Tree *) :: Bool where ...

Compute when a type is in a tree.

Equations

In x (Leaf y) = x == y

In x (Node l r) = In x l || In x r

type family Contents mm :: Tree *

Type family that represents the contents of a map.

Lookup

lookup :: (MapsTo k v mm, Ord k, Show k, Typeable k, Typeable v) => k -> mm -> Either Error v

Lookup a key in the map.

lookupM :: (MapsTo k v mm, Ord k, Show k, Typeable k, Typeable v) => k -> mm -> CompilerM v

Monadic version of lookup.

Extraction of maps, keys and values

keys :: forall k v mm. MapsTo k v mm => mm -> [k]

Get all the keys of type k that are associated to values of type v in the given map.

values :: forall k v mm. MapsTo k v mm => mm -> [v]

Get all the values of type v that are associated to keys of type k in the given map.

innerMap :: MapsTo k v mm => mm -> Map k v

Get the inner map.

Combination of maps

data a :& b infixr 5

Combinator for maps.

Constructors

a :& b infixr 5

(<.+>) :: (MapsTo k v mm, Ord k) => Map k v -> mm -> mm infixl 4

Add a map. The addition is biased towards the new map. If the two maps have the same key, then the value of the right map is discarded, and the value of the left map is kept.

(<.++>) :: (Ord k, MapsTo k v mm) => [(k, v)] -> mm -> mm infixr 6

Like .+ but its left argument is a list instead of a map. Useful when we want to add a list to a composite map.

(.&) :: Ord k0 => [(k0, v0)] -> mm -> Map k0 v0 :& mm infixr 6

Combine a list of key values pairs with an existing map.

(.&.) :: (Ord k0, Ord k1) => [(k0, v0)] -> [(k1, v1)] -> Map k0 v0 :& Map k1 v1 infixr 6

Combine lists of key values pairs into a pair of maps.

replaceInnerMap :: MapsTo k v mm => mm -> Map k v -> mm

Replace the inner map with the given one.