TorXakis.Compiler.ValExpr.ExpDecl

Description

Compilation functions related to expression declarations.

Documentation

class HasVarReferences e where

Expressions that have variable references.

Minimal complete definition

mapRefToDecls

Methods

mapRefToDecls

Arguments

:: (MapsTo Text [Loc FuncDeclE] mm, MapsTo Text (Loc VarDeclE) mm)

=> mm
Predefined functions

-> e

-> CompilerM [(Loc VarRefE, Either (Loc VarDeclE) [Loc FuncDeclE])]

Map variable references to the entities they refer to.

NOTE: property to check:

the number of 'Loc FuncDeclE' entities in the function declaration should equal the length of the list returned by this function.

This ensures that the mapping returned is complete.

This class could be replaced by DefinesAMap which already includes such an invariant check.

mapRefToDecls

Arguments

:: (HasVarReferences e, MapsTo Text [Loc FuncDeclE] mm, MapsTo Text (Loc VarDeclE) mm)

=> mm
Predefined functions

-> e

-> CompilerM [(Loc VarRefE, Either (Loc VarDeclE) [Loc FuncDeclE])]

Map variable references to the entities they refer to.

NOTE: property to check:

the number of 'Loc FuncDeclE' entities in the function declaration should equal the length of the list returned by this function.

This ensures that the mapping returned is complete.

This class could be replaced by DefinesAMap which already includes such an invariant check.