TorXakis.Compiler.ValExpr.FuncId

Description

Compilation functions related to TorXakis function id's.

Documentation

cstrToAccFuncId

Arguments

:: MapsTo Text SortId mm

=> mm

-> CstrId
Id of the containing constructor

-> FieldDecl

-> CompilerM FuncId

Define the accessor function for a constructor, using the given field.

cstrToIsCstrFuncId :: CstrId -> CompilerM FuncId

Make a function id from the given constructor id.

sortFromStringFuncId :: SortId -> CompilerM FuncId

Make a FuncId for a string to sort function.

sortToStringFuncId :: SortId -> CompilerM FuncId

Make a FuncId for a sort to string function.

adtsToFuncIds :: MapsTo Text SortId mm => mm -> [ADTDecl] -> CompilerM [(Loc FuncDeclE, FuncId)]

Create the function id's that the ADT implicitly defines.

funcIdAsSignature :: FuncId -> Signature

Return the function id as a signature.

funcDeclsToFuncIds :: MapsTo Text SortId mm => mm -> [FuncDecl] -> CompilerM [(Loc FuncDeclE, FuncId)]

Compile a list of function declarations to a list of pairs consisting of the function location and its id.

getStdFuncIds :: CompilerM [(Loc FuncDeclE, FuncId)]

Generate the function id's that correspond to the functions in the standard function table (stdFuncTable)