TorXakis.Compiler.ValExpr.SortId

Description

Compilation functions related to to SortId values.

Sort lookups

sortIdOfVarDecl :: MapsTo Text SortId mm => mm -> VarDecl -> Either Error SortId

Find the SortId of a variable declaration.

sortIdOfVarDeclM :: MapsTo Text SortId mm => mm -> VarDecl -> CompilerM SortId

Monadic version of sortIdOfVarDecl.

checkSortIds :: SortId -> SortId -> Either Error ()

Check that the two sorts match.

sortIdConst :: Const -> Maybe SortId

SortId of a constant.

sortIds :: MapsTo Text SortId mm => mm -> [OfSort] -> CompilerM [SortId]

exitSortIds :: ExitSort -> [SortId]

offerSid :: (MapsTo Text SortId mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc FuncDeclE) Signature mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm) => mm -> ChanOfferDecl -> CompilerM SortId

exitSort :: (HasExitSorts e, MapsTo Text SortId mm, MapsTo (Loc ChanRefE) (Loc ChanDeclE) mm, MapsTo (Loc ChanDeclE) ChanId mm, MapsTo ProcId () mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm, MapsTo (Loc FuncDeclE) Signature mm) => mm -> e -> CompilerM ExitSort

Obtain the exit sorts for an expression.

Type (Sort) inference.

inferVarTypes :: HasTypedVars mm e => mm -> e -> CompilerM [(Loc VarDeclE, SortId)]

class HasTypedVars mm e where

An expression has typed-variables if a map can be found from the location of variable declarations, to their associated SortId.

Minimal complete definition

inferVarTypes

Methods

inferVarTypes :: mm -> e -> CompilerM [(Loc VarDeclE, SortId)]

inferTypes :: (MapsTo Text SortId mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc FuncDeclE) Signature mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm) => mm -> [FuncDecl] -> CompilerM (Map (Loc VarDeclE) SortId)

Infer the types in a list of function declarations.

inferExpTypes :: (MapsTo Text SortId mm, MapsTo (Loc VarDeclE) SortId mm, MapsTo (Loc FuncDeclE) Signature mm, MapsTo (Loc VarRefE) (Either (Loc VarDeclE) [Loc FuncDeclE]) mm) => mm -> ExpDecl -> Either Error [SortId]

Infer the type of an expression. Due to function overloading an expression could have multiple types, e.g.:

fromString("33")

Could be a TorXakis Int, String, Bool, or even an ADT.

Compilation

compileToSortId :: [ADTDecl] -> CompilerM (Map Text SortId)

Construct a list of sort ID's from a list of ADT declarations.

The ID of the ADT will coincide with the ID of the SortId.