TorXakis.Compiler.ValExpr.ValExpr

Description

Compilation functions related to TorXakis value expressions.

Documentation

expDeclToValExpr

Arguments

:: Map (Loc VarRefE) (Either VarId [(Signature, Handler VarId)])
Variable definitions associated to a given reference.

-> SortId
Expected SortId for the expression.

-> ExpDecl

-> Either Error (ValExpr VarId)

Compile an expression declaration into a value expression, given a map of variable references to the entities they refer, and the expected sort id.

parValDeclToMap :: Map (Loc VarRefE) (Either VarId [(Signature, Handler VarId)]) -> ParLetVarDecl -> Either Error (Map VarId (ValExpr VarId))

Compile a parallel let-variable declaration into a map of variable id's to the expressions they are associated with.