TorXakis.Parser

Description

Documentation

data ParsedDefs

TorXakis definitions generated by the parser.

adts :: Lens' ParsedDefs [ADTDecl]

funcs :: Lens' ParsedDefs [FuncDecl]

consts :: Lens' ParsedDefs [FuncDecl]

models :: Lens' ParsedDefs [ModelDecl]

chdecls :: Lens' ParsedDefs [ChanDecl]

procs :: Lens' ParsedDefs [ProcDecl]

txsP :: TxsParser ParsedDefs

Root parser for the TorXakis language.

parseFile :: FilePath -> IO (Either Error ParsedDefs)

Parse a TorXakis file from file.

parseString :: FilePath -> String -> Either Error ParsedDefs

Parse a TorXakis file from a string.

parse

Arguments

:: Int
Initial value for the unique id's counter

-> String
Name of the source from which the input was read.

-> String
Input for the parser

-> TxsParser a
Parser to run

-> Either Error a

Parse a string using the initial value for the unique id's counter.