TorXakis.Parser.Common

Description

Common types and functionality for the TorXakis parser.

Documentation

type TxsParser = ParsecT ParserInput St Identity

txsSymbol

Arguments

:: String
String representation of the symbol.

-> TxsParser ()

Parse given symbol, discarding its result.

identifier :: TxsParser Text

Parser for identifiers, which may start with lower or upper case letters.

mkLoc :: TxsParser (Loc t)

Make a new location while parsing.

lcIdentifier :: TxsParser Text

Parser for lower-case identifiers.

txsLexeme :: TxsParser a -> TxsParser a

Parser for lexemes.

ucIdentifier :: String -> TxsParser Text

Parser for upper-case identifiers.

txsIntP :: TxsParser Integer

Parser for integer.

txsStringP :: TxsParser Text

Parser for strings.

txsWhitespace :: TxsParser ()

Parser that ignores white-spaces.

inP :: TxsParser a -> TxsParser a

Parse expressions of the form "IN exp NI", where p is the expressions parser.

tryIdentifier :: TxsParser Text

Like identifier, but try.

declP

Arguments

:: String
Name of the declaration.

-> (Text -> Loc l -> TxsParser a)
Parser for the body of the declaration. The name given as a parameter is the name of the declaration. The location given as parameter is the location in which the declaration name was found.

-> TxsParser a

Parser for declarations of the form:

XDEF n ::= p ENDEF

Where n is the name of the definition, p is the parser that parses the body of the declaration, and XDEF is the declaration name (e.g. MODELDEF).

declWithParamsP

Arguments

:: String

-> TxsParser params

-> (params -> Text -> Loc l -> TxsParser a)

-> Bool
Is ENDEF expected?

-> TxsParser a

Parser for declarations with parameters of the form: XDEF n params ::= body ENDEF

Where n is the name of the definition, params are the parameters taken by the declaration, body is the body of the declaration, and XDEF is the declaration name (e.g. PROCDEF). The boolean parameter of this function indicates whether 'ENDDEF' is required.