TorXakis.Parser.Data

Description

Abstract syntax trees for the TorXakis parser and compiler.

Documentation

data St

State of the parser.

mkState :: Int -> St

Create anew state.

nodeLoc :: ParseTree t c -> Loc t

nodeType :: ParseTree t c -> t

incId :: St -> St

Increment the id of the state.

nextId :: St -> Int

data Loc t

Location associated to the elements being parsed.

Constructors

Loc

Fields

line :: Int

Line in which a declaration occurs.

start :: Int

Start column.

locUid :: Int

Unique identifier.

PredefLoc

Fields

locName :: Text

Name of the predefined location

locUid :: Int

Unique identifier.

ExtraAut Text (Loc t)

locFromLoc :: Loc t -> Loc u

Change extract the location of the metadata, and change its type from t to u. This is useful when defining parsed entities whose locations coincide, like expressions and variable-references or constant-literals.

class HasLoc a t | a -> t where

Class of entities that have a location associated to them.

Minimal complete definition

getLoc, setLoc

Methods

getLoc :: a -> Loc t

Get the location of the entity.

loc' :: Lens' a (Loc t)

Location lens.

class IsVariable v where

Entities that are variables.

Minimal complete definition

varName

Methods

varName :: v -> Text

Name of a variable.

Name

data Name t

Name of an entity.

toText :: Name t -> Text

Types of the entities.

data ADTE

ADT.

data CstrE

Constructor.

data FieldE

Field of a constructor.

data SortRefE

Reference to an existing (previously defined or primitive) sort.

data FuncDeclE

Function declaration.

data VarDeclE

A variable declaration.

data ExpDeclE

An expression

data VarRefE

A variable occurrence in an expression. It is assumed to be a **reference** to an existing variable.

data ChanDeclE

Channel declaration.

data ChanRefE

Channel reference.

data ProcDeclE

Process declaration.

data StautDeclE

State automaton declaration.

data StateDeclE

State declaration.

data StateRefE

Reference to a state.

data PurpDeclE

Purpose declaration.

data GoalDeclE

Goal declaration.

data CnectDeclE

Connect declaration.

data ActPrefE

Action prefix.

Declarations.

ADT's

type ADTDecl = ParseTree ADTE [CstrDecl]

mkADTDecl :: Text -> Loc ADTE -> [CstrDecl] -> ADTDecl

Make an ADT declaration.

adtName :: ADTDecl -> Text

Name of the ADT declaration.

constructors :: ADTDecl -> [CstrDecl]

Constructors of the ADT.

type CstrDecl = ParseTree CstrE [FieldDecl]

ADT constructor declaration.

mkCstrDecl :: Text -> Loc CstrE -> [FieldDecl] -> CstrDecl

Make a constructor declaration.

cstrName :: CstrDecl -> Text

Name of the constructor.

cstrFields :: CstrDecl -> [FieldDecl]

Fields of the constructor.

type FieldDecl = ParseTree FieldE OfSort

Field declarations.

mkFieldDecl :: Text -> Loc FieldE -> OfSort -> FieldDecl

Make a field declaration.

fieldName :: FieldDecl -> Text

Name of the field declaration.

fieldSort :: FieldDecl -> (Text, Loc SortRefE)

Get the field of a sort, and the metadata associated to it.

Type declarations

type OfSort = ParseTree SortRefE ()

Reference to an existing type

mkOfSort :: Text -> Loc SortRefE -> OfSort

Make a declaration of an entity belonging to a sort.

sortRefName :: OfSort -> Text

Name of the sort reference.

Functions

type FuncDecl = ParseTree FuncDeclE FuncComps

Function declarations.

mkFuncDecl :: Text -> Loc FuncDeclE -> [VarDecl] -> OfSort -> ExpDecl -> FuncDecl

Make a function declaration.

funcName :: FuncDecl -> Text

Name of the function declaration.

funcParams :: FuncDecl -> [VarDecl]

Function parameters.

funcBody :: FuncDecl -> ExpDecl

Function body,

funcRetSort :: FuncDecl -> (Text, Loc SortRefE)

Return sort of a function.

type VarDecl = ParseTree VarDeclE OfSort

Variable declarations (with an explicit sort).

mkVarDecl :: Text -> Loc VarDeclE -> OfSort -> VarDecl

Make a variable declaration.

varDeclSort :: VarDecl -> (Text, Loc SortRefE)

Sort of a variable declaration.

type IVarDecl = ParseTree VarDeclE (Maybe OfSort)

Implicit variable declaration (maybe with a sort associated to it).

mkIVarDecl :: Text -> Loc VarDeclE -> Maybe OfSort -> IVarDecl

Make an implicit variable declaration, which might have an optional type associated to it.

ivarDeclSort :: IVarDecl -> Maybe OfSort

type VarRef = ParseTree VarRefE ()

Variable references.

mkVarRef :: Text -> Loc VarRefE -> VarRef

Make a variable reference.

varName :: IsVariable v => v -> Text

Name of a variable.

Expressions

type ExpDecl = ParseTree ExpDeclE ExpChild

Declaration of an expression.

data ExpChild

Constructors

VarRef (Name VarRefE) (Loc VarRefE)

ConstLit Const

LetExp [ParLetVarDecl] ExpDecl
A let expression allows to introduce a series of value bindings of the form:

x0 = v0, ..., xn = vn

A let expression contains a list of lists that have the form above. Values declared within one '[LetVarDecl]' list introduce variable names in parallel (in the sense that one variable in the list cannot be used in the expressions within that list). However, values declared in '[LetVarDecl]' lists can be used in subsequent '[LetVarDecl]' lists. For instance, the following let expression:

LET x = 1, y = 5; z = x + y IN ...

Will be parsed to the following list

[[(x, 1), (y, 5)], [(z, x + y)]]

Here x and y cannot be used in the expressions of the first list, but it can be used in the expressions of the second.

If ExpDecl ExpDecl ExpDecl

Fappl (Name VarRefE) (Loc VarRefE) [ExpDecl]
Function application. A function is applied to a list of expressions.

data Const

Constant declarations.

Constructors

BoolConst Bool

IntConst Integer

StringConst Text

RegexConst Text

AnyConst

expChild :: ExpDecl -> ExpChild

Child of an expression.

childExps :: ExpDecl -> [ExpDecl]

Get the child-expressions of an expression.

mkVarExp :: Loc ExpDeclE -> Text -> ExpDecl

Make a variable expression. The location of the expression will become the location of the variable.

mkBoolConstExp :: Loc ExpDeclE -> Bool -> ExpDecl

Make a Boolean constant expression.

mkIntConstExp :: Loc ExpDeclE -> Integer -> ExpDecl

Make an integer constant expression.

mkStringConstExp :: Loc ExpDeclE -> Text -> ExpDecl

Make a string constant expression.

mkRegexConstExp :: Loc ExpDeclE -> Text -> ExpDecl

Make a regex constant expression.

mkLetExpDecl :: [ParLetVarDecl] -> ExpDecl -> Loc ExpDeclE -> ExpDecl

Make a let-expression declaration.

mkITEExpDecl :: Loc ExpDeclE -> ExpDecl -> ExpDecl -> ExpDecl -> ExpDecl

Make an if-then-else expression declaration.

mkFappl :: Loc ExpDeclE -> Loc VarRefE -> Text -> [ExpDecl] -> ExpDecl

Make a function-application expression declaration.

mkLetVarDecl :: Text -> Maybe OfSort -> ExpDecl -> Loc VarDeclE -> LetVarDecl

Make a let variable declaration.

expLetVarDecls :: ExpDecl -> [[LetVarDecl]]

Extract all the let-variable declarations of an expression.

type LetVarDecl = ParseTree VarDeclE (Maybe OfSort, ExpDecl)

Let-variable declarations.

varDeclExp :: LetVarDecl -> ExpDecl

Expression declaration in a let-variable declaration.

letVarDeclSortName :: LetVarDecl -> Maybe (Text, Loc SortRefE)

Sort name of a let-variable declaration.

newtype ParLetVarDecl

Parallel let-variable declarations.

Constructors

ParLetVarDecl [LetVarDecl]

Models

type ModelDecl = ParseTree ModelDeclE ModelComps

Model declaration.

mkModelDecl

Arguments

:: Text
Model name.

-> Loc ModelDeclE
Location of the model.

-> [ChanRef]
References to input channels.

-> [ChanRef]
References to output channels.

-> Maybe [Set ChanRef]
References to sets of synchronized channels.

-> BExpDecl
Behavior expression that defines the model.

-> ModelDecl

Make a model declaration.

modelName :: ModelDecl -> Text

Name of the model.

modelIns :: ModelDecl -> [ChanRef]

Model input channels.

modelOuts :: ModelDecl -> [ChanRef]

Model output channels.

modelSyncs :: ModelDecl -> Maybe [Set ChanRef]

Model synchronization channels.

modelBExp :: ModelDecl -> BExpDecl

Model behavior expressions.

data BExpDecl

Behavior expression declaration.

Constructors

Stop
STOP operator.

ActPref (Loc ActPrefE) ActOfferDecl BExpDecl
>-> (action prefix) operator.

LetBExp [ParLetVarDecl] BExpDecl
LET declarations for behavior expressions.

Pappl (Name ProcRefE) (Loc ProcRefE) [ChanRef] [ExpDecl]
Process instantiation.

Par (Loc ParOpE) SyncOn BExpDecl BExpDecl
Parallel operators.

Enable (Loc EnableE) BExpDecl BExpDecl
Enable operator.

Accept (Loc AcceptE) [ChanOfferDecl] BExpDecl
ACCEPT operator.

Note that while the parser will allow ACCEPTs in arbitrary positions, the compiler will check that they only occur after an enable operator (>>>)

Disable (Loc DisableE) BExpDecl BExpDecl
Disable operator.

Interrupt (Loc InterruptE) BExpDecl BExpDecl
Interrupt operator.

Choice (Loc ChoiceE) BExpDecl BExpDecl
Choice operator.

Guard ExpDecl BExpDecl
Guard operator.

Hide (Loc HideE) [ChanDecl] BExpDecl
Hide operator.

data ActOfferDecl

Action offer declaration.

Constructors

ActOfferDecl

Fields

_offers :: [OfferDecl]

_constraint :: Maybe ExpDecl

data OfferDecl

Offer declaration.

Constructors

OfferDecl ChanRef [ChanOfferDecl]

data ChanOfferDecl

Channel offer declarations.

Note that a receiving action with an explicit type declaration are only needed to simplify the type inference of exit variables used in expressions of the form 'EXIT ? v :: T'.

Constructors

QuestD IVarDecl

ExclD ExpDecl

data SyncOn

Channels to sync on in a parallel operator.

Constructors

All
Sync on all channels, this is the result of parsing ||

OnlyOn [ChanRef]
Sync only on the given channels. This is the result of parsing either ||| or '|[...]|'. Parsing ||| will result in an empty list, meaning that full interleaving is allowed.

chanOfferIvarDecl :: ChanOfferDecl -> Maybe IVarDecl

Implicit variable (if any) of a channel offer declaration.

actOfferDecls :: ActOfferDecl -> [IVarDecl]

Get all the variable declarations introduced by receiving actions of the form 'Ch ? v'.

offerDecls :: OfferDecl -> [IVarDecl]

Get the implicit variable declarations of an offer declaration.

chanRefOfOfferDecl :: OfferDecl -> ChanRef

Reference to a channel in an offer declaration.

chanOfferDecls :: ChanOfferDecl -> [IVarDecl]

Implicit variable declarations of a channel offer.

asVarReflLoc :: Loc VarDeclE -> Loc VarRefE

Transform the location of a variable declaration into the location of a variable reference. A variable declaration refers to itself.

We do not want to export locFromLoc since we will be loosing the type-guarantees offered by the phantom type. By having locFromLoc private to this module we are in control of which conversions are allowed.

Channels

type ChanDecl = ParseTree ChanDeclE [OfSort]

Channel declaration.

mkChanDecl :: Text -> Loc ChanDeclE -> [OfSort] -> ChanDecl

Make a channel declaration.

chanDeclName :: ChanDecl -> Text

Name of the channel declaration.

chanDeclSorts :: ChanDecl -> [(Text, Loc SortRefE)]

Sorts of a channel declaration with their locations, which can be used for lookup and error reporting.

type ChanRef = ParseTree ChanRefE ()

Channel references.

mkChanRef

Arguments

:: Text
Name of the channel that is being referred.

-> Loc ChanRefE
Location where the reference took place.

-> ChanRef

Make a channel reference.

chanRefName :: ChanRef -> Text

Name of the channel reference.

Processes

type ProcDecl = ParseTree ProcDeclE ProcComps

Process declaration.

mkProcDecl :: Text -> Loc ProcDeclE -> [ChanDecl] -> [VarDecl] -> ExitSortDecl -> BExpDecl -> ProcDecl

Make a process declaration.

procDeclName :: ProcDecl -> Text

Name of a process declaration.

procDeclComps :: ProcDecl -> ProcComps

Components of a process declaration.

procDeclBody :: ProcDecl -> BExpDecl

Body of a process declaration.

procDeclRetSort :: ProcDecl -> ExitSortDecl

Return sort of a process declaration.

procDeclChParams :: ProcDecl -> [ChanDecl]

Parameters of a process declaration.

procDeclParams :: ProcDecl -> [VarDecl]

Parameters of a process declaration.

data ProcComps

Components of a process.

procChParams :: ProcComps -> [ChanDecl]

procParams :: ProcComps -> [VarDecl]

procRetSort :: ProcComps -> ExitSortDecl

procBody :: ProcComps -> BExpDecl

procRefName :: Text -> Name ProcRefE

Process reference name.

data ExitSortDecl

Possible exit sorts of a process.

Constructors

NoExitD

ExitD [OfSort]

HitD

State automata

type StautDecl = ParseTree StautDeclE StautComps

State automaton declaration.

data StautItem

Item of a state automaton.

Constructors

States [StateDecl]

StVarDecl [VarDecl]

InitState InitStateDecl

Trans [Transition]

type StateDecl = ParseTree StateDeclE ()

Declaration of an automaton state.

type StateRef = ParseTree StateRefE ()

Reference to a previously declared automaton state.

data StUpdate

State automaton update.

Constructors

StUpdate [VarRef] ExpDecl

data Transition

State automaton transition.

Constructors

Transition StateRef ActOfferDecl [StUpdate] StateRef

mkStautDecl :: Text -> Loc StautDeclE -> [ChanDecl] -> [VarDecl] -> ExitSortDecl -> [StautItem] -> StautDecl

Make a state automaton declaration.

mkStateDecl :: Text -> Loc StateDeclE -> StateDecl

Make a state declaration.

stateDeclName :: StateDecl -> Text

Name the state declaration.

mkStateRef :: Text -> Loc StateRefE -> StateRef

Make a state reference.

stateRefName :: StateRef -> Text

Name of the state reference.

stautName :: StautDecl -> Text

Name of the state automaton.

asProcDeclLoc :: StautDecl -> Loc ProcDeclE

Return the location of a state automaton as the location of a process declaration.

stautDeclChParams :: StautDecl -> [ChanDecl]

Get the channel declarations of a state automaton.

stautDeclParams :: StautDecl -> [VarDecl]

Get the formal parameters of a state automaton.

stautDeclRetSort :: StautDecl -> ExitSortDecl

Get the return sort of a state automaton.

stautDeclComps :: StautDecl -> [StautItem]

Get the components of a state automaton.

stautDeclStates :: StautDecl -> [StateDecl]

Extract the states declared in the automaton.

stautDeclInnerVars :: StautDecl -> [VarDecl]

Extract the variables declared in the automaton.

stautInitStates :: StautDecl -> [InitStateDecl]

Extract the initial states declared in the automaton.

stautTrans :: StautDecl -> [Transition]

Extract the transitions declared in the automaton.

data InitStateDecl

Declaration of an initial state.

Constructors

InitStateDecl StateRef [StUpdate]

mkInitState :: StateRef -> [StUpdate] -> StautItem

Make an initial state declaration.

Purposes

type PurpDecl = ParseTree PurpDeclE PurpComps

Purpose declaration.

mkPurpDecl :: Text -> Loc PurpDeclE -> [ChanRef] -> [ChanRef] -> Maybe [Set ChanRef] -> [TestGoalDecl] -> PurpDecl

Make a purpose declaration.

data PurpComps

Components of a purpose.

Constructors

PurpComps

Fields

purpIns :: [ChanRef]

purpOuts :: [ChanRef]

purpSyncs :: Maybe [Set ChanRef]

goals :: [TestGoalDecl]

purpDeclName :: PurpDecl -> Text

Name of the purpose declaration.

purpDeclIns :: PurpDecl -> [ChanRef]

Input channels of a purpose declaration.

purpDeclOuts :: PurpDecl -> [ChanRef]

Output channels of a purpose declaration.

purpDeclSyncs :: PurpDecl -> Maybe [Set ChanRef]

Synchronization channels of a purpose declaration.

purpDeclGoals :: PurpDecl -> [TestGoalDecl]

Goals of a purpose declaration.

type TestGoalDecl = ParseTree GoalDeclE BExpDecl

Declaration of a test goal.

mkTestGoalDecl :: Text -> Loc GoalDeclE -> BExpDecl -> TestGoalDecl

Make a test goal declaration.

testGoalDeclName :: TestGoalDecl -> Text

Name of a test goal declaration.

testGoalDeclBExp :: TestGoalDecl -> BExpDecl

Behavior expression of a test goal declaration.

Connections

type CnectDecl = ParseTree CnectDeclE (CnectType, [CnectItem], [CodecItem])

CNECTDEF declaration.

data CnectType

Connect type declaration.

Constructors

CTClient

CTServer

data CnectItem

Connect item declaration.

Constructors

CnectItem

Fields

cnectCh :: ChanRef

cnectType :: CnectItemType

host :: Text

port :: Integer

data CnectItemType

Type of the connect item.

Constructors

ChanIn

ChanOut

data CodecItem

Codec item declaration.

Constructors

CodecItem

Fields

codecOffer :: OfferDecl

codecChOffer :: ChanOfferDecl

codecType :: CodecType

data CodecType

Type of the codec item.

Constructors

Decode

Encode

mkCnectDecl :: Text -> Loc CnectDeclE -> CnectType -> [CnectItem] -> [CodecItem] -> CnectDecl

Make a connect declaration.

cnectDeclName :: CnectDecl -> Text

Name of the connect declaration.

cnectDeclType :: CnectDecl -> CnectType

Type of the connect declaration.

cnectDeclCnectItems :: CnectDecl -> [CnectItem]

Connect items of the connect declaration.

cnectDeclCodecs :: CnectDecl -> [CodecItem]

Codec items of the connect declaration.

Mappers

type MapperDecl = ParseTree MapperDeclE ModelComps

Mapper declaration.

data MapperDeclE

Mapper declaration.

mkMapperDecl :: Text -> Loc MapperDeclE -> [ChanRef] -> [ChanRef] -> Maybe [Set ChanRef] -> BExpDecl -> MapperDecl

Make a mapper declaration.

mapperName :: MapperDecl -> Text

Name of the mapper.

mapperIns :: MapperDecl -> [ChanRef]

Input channels of the mapper declaration.

mapperOuts :: MapperDecl -> [ChanRef]

Output channels of the mapper declaration.

mapperSyncs :: MapperDecl -> Maybe [Set ChanRef]

Synchronization channels of the mapper declaration.

mapperBExp :: MapperDecl -> BExpDecl

Behavior expression of the mapper declaration.

Location of the entities

getLoc :: HasLoc a t => a -> Loc t

Get the location of the entity.

loc' :: HasLoc a t => Lens' a (Loc t)

Location lens.

Parsed definitions

data ParsedDefs

TorXakis definitions generated by the parser.

adts :: Lens' ParsedDefs [ADTDecl]

funcs :: Lens' ParsedDefs [FuncDecl]

consts :: Lens' ParsedDefs [FuncDecl]

models :: Lens' ParsedDefs [ModelDecl]

chdecls :: Lens' ParsedDefs [ChanDecl]

procs :: Lens' ParsedDefs [ProcDecl]

emptyPds :: ParsedDefs

Empty parsed definitions.

stauts :: Lens' ParsedDefs [StautDecl]

purps :: Lens' ParsedDefs [PurpDecl]

cnects :: Lens' ParsedDefs [CnectDecl]

mappers :: Lens' ParsedDefs [MapperDecl]