TorXakis.Parser.TypeDefs

Description

Parser for type declarations (TorXakis ADT's)

Documentation

adtP :: TxsParser ADTDecl

Parser for type declarations (TorXakis ADT's).

sortP :: TxsParser OfSort

Parser for Sorts.

idOfSortsP

Arguments

:: String
Start symbol for the fields declaration.

-> String
End symbol for the fields declaration.

-> (Text -> Loc t -> OfSort -> d)

-> TxsParser [d]

Parser of a list of identifiers declarations, with an associated sort. The list of identifiers is delimited by the given start and end symbols, and separated by a semi-colon.

identifiers :: Sort

ofSortP :: TxsParser OfSort

Parser for the declaration of a sort.